requirements.txt
Django==3.2
celery==5.2.7
djangorestframework==3.14.0
django-filter==22.1

# Persistent Store
# -------------------------------------------------
psycopg2-binary==2.9.5
redis==4.3.4
dj-database-url==0.5.0

# Extensions
# -------------------------------------------------
django-extensions==3.2.1
django-sampledatahelper==0.5
unicode-slugify==0.1.5
XlsxWriter==3.0.3
arrow==1.2.3
hashids==1.3.1

# Django Model Helpers
# -------------------------------------------------
django-uuid-upload-path==1.0.0
Pillow==9.3.0

# Auth
# -------------------------------------------------
django-allauth==0.51.0

# Admin
# -------------------------------------------------
django_simple_history==3.2.0

# Frontend Helpers
# -------------------------------------------------
django-bootstrap3==22.1
django-bootstrap-breadcrumbs==0.9.2

# Markdown Editor
# -------------------------------------------------
django-pagedown==2.2.1
django-markdownx==3.0.1
markdown2==2.4.6

# Ticket

qrcode==7.3.1


#server
#uwsgi
